+++ optimus/conf/registry.py
@@ -20,8 +20,6 @@ except KeyError:
     raise ImportError(("Project cannot be imported, because "
                        "environment variable {} is "
                        "undefined.").format(PROJECT_DIR_ENVVAR))
-else:
-    print("🧐 os.environ[PROJECT_DIR_ENVVAR]=", os.environ[PROJECT_DIR_ENVVAR])
 
 try:
     name = os.environ[SETTINGS_NAME_ENVVAR]
@@ -32,10 +30,6 @@ except KeyError:
     raise ImportError(("Settings cannot be imported, because "
                        "environment variable {} is "
                        "undefined.").format(SETTINGS_NAME_ENVVAR))
-else:
-    print("🧐 os.environ[SETTINGS_NAME_ENVVAR]=", os.environ[SETTINGS_NAME_ENVVAR])
 
 # Reachable settings object
-print("🧐 import_settings(", name, basedir, ")")
 settings = import_settings(name, basedir)
-print("🧐 do you see me ?")